Add tests for counting comments
