+++ src/countComments.js
@@ -1,0 +1,5 @@
+function countComments(comments) {
+  return comments.length;
+}
+
+export { countComments as default };
