--- src/test/comments.test.js
@@ -1,7 +1,0 @@
-import { countComments } from "../comments.js";
-
-describe("Testing comments counter", () => {
-  test("First test", () => {
-    expect(countComments([{}, {}, {}])).toBe(3);
-  });
-});
